Update cmake to handle MSYS correctly.
Project 0 2D v0.03:
 -> Remove cbp project coz its not useful anymore
 -> update cmake to copy config files
 -> print a msg when a config file is not found

+++ P0/include/Player_Base.hh
@@ -36,7 +36,7 @@ class Player_Base : public Character_Base
 		//animation variables
 		int frame, arrow_frame, move_status;
 
-		int DeltaTicks;
+		long DeltaTicks;
 
 		/***Arrow***/
 		RGBSurface Arrow_Tile;
@@ -110,11 +110,11 @@ class Player_Base : public Character_Base
             return moving_status;
         }
 
-		inline void Set_DeltaTicks(int new_DeltaTicks)
+		inline void Set_DeltaTicks(long new_DeltaTicks)
         {
             DeltaTicks = new_DeltaTicks;
         }
-        inline int Get_DeltaTicks() const
+        inline long Get_DeltaTicks() const
         {
             return DeltaTicks;
         }
